refactor(git): Git Generator absorbed into main generator
